backend/services/order_service/requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
dj-database-url==2.1.0
django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.27.2
gunicorn==22.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pika==1.3.2
psycopg2-binary==2.9.10
pyjwt==2.12.1
python-dotenv==1.2.2
pyyaml==6.0.3
redis==5.0.4
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
sqlparse==0.5.5
typing-extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
